fin quete 24 route

+++ src/Controller/ProgramController.php
@@ -8,6 +8,7 @@ use App\Entity\Season;
 use App\Form\ProgramType;
 use App\Repository\ProgramRepository;
 use App\Service\ProgramDuration;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -17,7 +18,7 @@ use Symfony\Component\Mime\Address;
 use Symfony\Component\Mime\Email;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\String\Slugger\SluggerInterface;
-
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 #[Route('/program', name: 'app_program_')]
 class ProgramController extends AbstractController
@@ -28,12 +29,13 @@ class ProgramController extends AbstractController
         return $this->render('program/index.html.twig', ['programs' => $programRepository->findAll()]);
     }
 
+    #[IsGranted('ROLE_CONTRIBUTOR')]
     #[Route('/new', name: 'new')]
     public function new(Request $request, MailerInterface $mailer, ProgramRepository $programRepository, SluggerInterface $slugger): Response
     {
         // Create a new Category Object
         $program = new Program();
-
+        $program->setOwner($this->getUser());
         // Create the form, linked with $category
         $form = $this->createForm(ProgramType::class, $program);
 
@@ -98,9 +100,16 @@ class ProgramController extends AbstractController
         return $this->render('program/episode_show.html.twig', ['program' => $program, 'season' => $season, 'episode' => $episode]);
     }
 
-    #[Route('/{program}/edit', requirements: ['program' => '\d+'],  name: 'edit', methods: ['GET', 'POST'])]
+    #[IsGranted('ROLE_CONTRIBUTOR')]
+    #[Route('/{slug}/edit', name: 'edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Program $program, programRepository $programRepository, SluggerInterface $slugger): Response
     {
+        // Check wether the logged in user is the owner of the program
+        if ($this->getUser() !== $program->getOwner()) {
+            // If not the owner, throws a 403 Access Denied exception
+            throw $this->createAccessDeniedException('Seul l\'auteur de cette série peut la modifier !');
+        }
+
         $form = $this->createForm(ProgramType::class, $program);
         $form->handleRequest($request);
 
@@ -123,9 +132,16 @@ class ProgramController extends AbstractController
         ]);
     }
 
+    #[IsGranted('ROLE_CONTRIBUTOR')]
     #[Route('/{id}', name: 'delete', methods: ['POST'])]
     public function delete(Request $request, Program $program, ProgramRepository $programRepository): Response
     {
+        // Check wether the logged in user is the owner of the program
+        if ($this->getUser() !== $program->getOwner()) {
+            // If not the owner, throws a 403 Access Denied exception
+            throw $this->createAccessDeniedException('Seul l\'auteur de cette série peut la supprimer !');
+        }
+
         if ($this->isCsrfTokenValid('delete' . $program->getId(), $request->request->get('_token'))) {
             $programRepository->remove($program, true);
             $this->addFlash(
